# docs/reference-v7.0.0/docs/06_APIs/openapi/motorsports-events-api-v1.yaml
openapi: 3.1.0
info:
  title: Motorsports Events API
  version: 6.1.0
  description: API canonique indépendante des fournisseurs.
servers:
  - url: /api/v1
paths:
...
  /events:
    get:
      operationId: listEvents
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - {name: disciplineId, in: query, schema: {type: string}}
        - {name: championshipId, in: query, schema: {type: string}}
        - {name: categoryId, in: query, schema: {type: [string, 'null']}}
        - {name: seasonId, in: query, schema: {type: string}}
        - {name: from, in: query, schema: {type: string, format: date-time}}
        - {name: to, in: query, schema: {type: string, format: date-time}}
      responses:
        '200': {description: Liste, content: {application/json: {schema: {$ref: '#/components/schemas/EventPage'}}}}
  /events/{eventId}:
    get:
      operationId: getEvent
      parameters: [{$ref: '#/components/parameters/EventId'}]
      responses:
        '200': {description: Épreuve, content: {application/json: {schema: {$ref: '#/components/schemas/Event'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
    EventId: {name: eventId, in: path, required: true, schema: {type: string}}
...
  responses:
    NotFound: {description: Ressource absente, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
...
  schemas:
...
    Event:
      type: object
      required: [id, seasonId, name, status]
      properties:
        id: {type: string}
        seasonId: {type: string}
        name: {type: string}
        startsAt: {type: [string, 'null'], format: date-time}
        endsAt: {type: [string, 'null'], format: date-time}
        status: {type: string}
        origin: {type: string, enum: [manual, provider, mixed]}
...
    PageMetadata:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: {type: integer}
        pageSize: {type: integer}
        totalItems: {type: integer}
        totalPages: {type: integer}
...
    EventPage: {allOf: [{$ref: '#/components/schemas/PageMetadata'}, {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Event'}}}}]}
...
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties: {code: {type: string}, message: {type: string}, details: {type: array, items: {type: object}}}